# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurante",
        "version": "1.0.0"
    },
    "paths": {
        "/api/entrada/simple/{sku}": {
            "put": {
                "tags": [
                    "Entrada de un producto simple"
                ],
                "description": "Entrada de un producto simple",
                "operationId": "App\\Http\\Controllers\\Movimiento\\Simple\\EntradaPutController::__invoke",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "ID of producto simple",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2387983d-fbd6-410f-8c71-dc92c4899d01"
                    },
                    {
                        "name": "cantidad",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "costo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/productoSimple": {
            "post": {
                "tags": [
                    "crear un nuevo producto simple"
                ],
                "summary": "crear un nuevo producto simple",
                "description": "Nuevo producto simple",
                "operationId": "App\\Http\\Controllers\\ProductoSimple\\ProductoSimplePostController::__invoke",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "costo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "precio",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}